better solver

+++ include/point.h
@@ -4,13 +4,13 @@
 class point {
 
 public:
-    int8_t x;
-    int8_t y;
-    point(int8_t xx = 0, int8_t yy = 0);
+    int16_t x;
+    int16_t y;
+    point(int16_t xx = 0, int16_t yy = 0);
     point(const point &a);
     point operator+(const point &a)const;
     point operator-(const point &a)const;
-    point operator*(int8_t m) const;
+    point operator*(int16_t m) const;
     point& operator=(const point &a);
     bool operator==(const point &a)const;
     bool operator!=(const point &a)const;
